ai_lms/requirements.txt
Django==5.1.4
mysqlclient==2.2.6
google-generativeai==0.8.3
Pillow==11.0.0
python-decouple==3.8
django-crispy-forms
crispy-bootstrap5
django-widget-tweaks==1.5.0
matplotlib==3.9.3
pandas==2.2.3
numpy
